utils/helpers.py
import yaml
import os
from pathlib import Path

def load_config():
    """Load configuration from YAML file"""
    config_path = Path(__file__).parent.parent / 'config' / 'config.yaml'
    with open(config_path, 'r') as file:
        return yaml.safe_load(file)

def ensure_reports_directory():
    """Ensure reports directory exists"""
    reports_dir = Path(__file__).parent.parent / 'reports'
    reports_dir.mkdir(exist_ok=True)
    return reports_dir
