requirements.txt
locust==2.24.0
PyYAML==6.0.1
python-dotenv==1.0.0
